fix: branch & company phone validation

+++ app/Http/Controllers/Api/BranchController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\BranchRequest;
 use App\Models\Branch;
 use Carbon\Carbon;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Yajra\DataTables\Facades\DataTables;
 
@@ -14,7 +15,9 @@ class BranchController extends Controller
 {
     function datatable()
     {
-        $data = Branch::query();
+        $user = Auth::user();
+        $data = Branch::query()->select('id', 'name', 'phone', 'address', 'main_branch', 'status')
+        ->filterByCompany($user->company_id);
 
         return DataTables::of($data)->addIndexColumn()->toJson();
     }
@@ -23,15 +26,24 @@ class BranchController extends Controller
     {
         DB::beginTransaction();
         try {
+            $user = Auth::user();
             $date = Carbon::now()->timezone(config('app.timezone'));
 
             $branch                 = new Branch();
+            $branch->company_id     = $user->company_id;
             $branch->name           = $request->name;
             $branch->phone          = $request->phone;
             $branch->address        = $request->address;
+            $branch->status         = Branch::STATUS_ACTIVE;
+            $branch->main_branch    = $request->is_main;
             $branch->created_at     = $date;
             $branch->updated_at     = $date;
             $branch->save();
+
+            if($request->is_main == Branch::IS_MAIN){ // jika request yg baru sudah menjadi main branch, maka yg lain harus di keluarkan dari main branch
+                $newId = $branch->id;
+                Branch::filterNotById($newId)->update(['main_branch' => Branch::IS_NOT_MAIN]);
+            }
 
             DB::commit();
 
@@ -46,11 +58,20 @@ class BranchController extends Controller
 
     function detail(Branch $branch)
     {
-        return ResponseFormatter::success($branch->makeHidden(['id', 'created_at', 'updated_at', 'deleted_at']));
+        $user = Auth::user();
+        if($user->company_id != $branch->company_id){
+            return ResponseFormatter::error(__('message.unauthorized'), ResponseFormatter::$errorUnauthorized);
+        }
+        return ResponseFormatter::success($branch->makeHidden(['id', 'created_at', 'updated_at', 'deleted_at', 'company_id']));
     }
 
     function update(BranchRequest $request, Branch $branch)
     {
+        $user = Auth::user();
+        if($user->company_id != $branch->company_id){
+            return ResponseFormatter::error(__('message.unauthorized'), ResponseFormatter::$errorUnauthorized);
+        }
+
         DB::beginTransaction();
         try {
             $date = Carbon::now()->timezone(config('app.timezone'));
@@ -58,6 +79,8 @@ class BranchController extends Controller
             $branch->name           = $request->name;
             $branch->phone          = $request->phone;
             $branch->address        = $request->address;
+            $branch->status         = $request->status;
+            $branch->main_branch    = $request->is_main;
             $branch->updated_at     = $date;
             $branch->save();
 
@@ -74,6 +97,11 @@ class BranchController extends Controller
 
     function delete(Branch $branch)
     {
+        $user = Auth::user();
+        if($user->company_id != $branch->company_id){
+            return ResponseFormatter::error(__('message.unauthorized'), ResponseFormatter::$errorUnauthorized);
+        }
+
         DB::beginTransaction();
         try {
             $branch->delete();
